src/test/java/com/uncledemy/salesmanagementsystem/salesserviceimpl/SalesServiceImplTest.java
package com.uncledemy.salesmanagementsystem.salesserviceimpl;

import com.uncledemy.salesmanagementsystem.dto.SalesDto;
import com.uncledemy.salesmanagementsystem.dto.SalesUpdateDto;
import com.uncledemy.salesmanagementsystem.exception.SalesManagementException;
import com.uncledemy.salesmanagementsystem.model.Sales;
import com.uncledemy.salesmanagementsystem.model.User;
import com.uncledemy.salesmanagementsystem.repository.ClientRepository;
import com.uncledemy.salesmanagementsystem.repository.ProductRepository;
import com.uncledemy.salesmanagementsystem.repository.SalesRepository;
import com.uncledemy.salesmanagementsystem.repository.UserRepository;
import com.uncledemy.salesmanagementsystem.service.SalesServiceImpl;
import org.junit.jupiter.api.BeforeEach;
import org.junit.jupiter.api.Test;
import org.mockito.InjectMocks;
import org.mockito.Mock;
import org.mockito.MockitoAnnotations;
import org.slf4j.Logger;
import org.slf4j.LoggerFactory;

import java.math.BigDecimal;
import java.util.HashMap;
import java.util.Map;

import static org.junit.jupiter.api.Assertions.assertNotNull;
import static org.junit.jupiter.api.Assertions.assertTrue;
import static org.mockito.Mockito.when;

class SalesServiceImplTest {

    @Mock
    private SalesRepository salesRepository;

    @Mock
    private ProductRepository productRepository;

    @Mock
    private UserRepository userRepository;

    @Mock
    private ClientRepository clientRepository;

    @InjectMocks
    private SalesServiceImpl salesService;

    private final Logger logger = LoggerFactory.getLogger(SalesServiceImpl.class);

    @BeforeEach
    void setUp() {
        MockitoAnnotations.initMocks(this);
    }

    @Test
    void createSales_ValidUserIdAndSalesDto_Success() throws SalesManagementException {
        long userId = 1L;
        SalesDto salesDto = new SalesDto();
        salesDto.setPhoneNumber("1234567890");
        Map<Long, Integer> productQuantities = new HashMap<>();
        productQuantities.put(1L, 1);
        salesDto.setProductQuantities(productQuantities);
        User user = new User();
        user.setId(userId);
        when(userRepository.findById(userId)).thenReturn(java.util.Optional.of(user));
        assertTrue(salesService.createSales(userId, salesDto));
    }

    @Test
    void updateSales_ValidSalesIdAndUserIdAndSalesUpdateDto_Success() throws SalesManagementException {
        long salesId = 1L;
        SalesUpdateDto salesUpdateDto = new SalesUpdateDto();
        salesUpdateDto.setQuantity(2);
        salesUpdateDto.setTotalAmount(BigDecimal.valueOf(200));
        Sales sales = new Sales();
        sales.setId(salesId);
        when(salesRepository.findById(salesId)).thenReturn(java.util.Optional.of(sales));
        assertTrue(salesService.updateSales(salesId, 1L, salesUpdateDto));
    }

    @Test
    void findSalesById_ValidSalesId_ReturnsSales() throws SalesManagementException {
        long salesId = 1L;
        Sales sales = new Sales();
        sales.setId(salesId);
        when(salesRepository.findById(salesId)).thenReturn(java.util.Optional.of(sales));
        assertNotNull(salesService.findSalesById(salesId));
    }

    @Test
    void getAllSales_ReturnsListOfSales() {
        // Mock data setup
        // when(salesRepository.findAll()).thenReturn(...);
        // Call service method
        // Verify result
    }

    @Test
    void generateSalesReport_ValidDates_ReturnsSalesReportDto() {
        // Mock data setup
        // when(salesRepository.findAll()).thenReturn(...);
        // Call service method
        // Verify result
    }

    // Add more test cases for edge cases and other scenarios...

}
